:lipstick: Refactor Notification to fully comply with Material 3 design principles
