+++ app/src/desktopMain/kotlin/com/crosspaste/ui/NotificationHost.kt
@@ -13,6 +13,8 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.contentColorFor
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -25,8 +27,10 @@ import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.window.Popup
 import androidx.compose.ui.window.PopupProperties
 import com.crosspaste.notification.Message
+import com.crosspaste.notification.MessageType
 import com.crosspaste.notification.NotificationManager
 import com.crosspaste.ui.base.NotificationCard
+import com.crosspaste.ui.base.NotificationContentCard
 import com.crosspaste.ui.theme.AppUISize.medium
 import com.crosspaste.ui.theme.AppUISize.tiny
 import kotlinx.coroutines.delay
@@ -102,10 +106,25 @@ private fun NotificationItemWrapper(
         enter = slideInVertically(initialOffsetY = { -it }) + fadeIn(),
         exit = slideOutHorizontally(targetOffsetX = { it }) + fadeOut() + shrinkVertically(),
     ) {
+        val containerColor =
+            when (notification.messageType) {
+                MessageType.Error -> MaterialTheme.colorScheme.errorContainer
+                MessageType.Warning -> MaterialTheme.colorScheme.tertiaryContainer
+                MessageType.Success -> MaterialTheme.colorScheme.primaryContainer
+                MessageType.Info -> MaterialTheme.colorScheme.secondaryContainer
+            }
+
+        val contentColor = contentColorFor(containerColor)
+
         NotificationCard(
-            notification = notification,
-            onCancelTapped = { visibleState.targetState = false },
             modifier = Modifier.padding(bottom = tiny),
-        )
+            containerColor = containerColor,
+        ) {
+            NotificationContentCard(
+                notification = notification,
+                contentColor = contentColor,
+                onCancelTapped = { visibleState.targetState = false },
+            )
+        }
     }
 }
